SUBMARINE-1207. workbench/rest to rest/workbench
