+++ submarine-server/server-core/src/test/java/org/apache/submarine/server/rest/workbecnh/SysDictItemRestApiTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.submarine.server.workbench.rest;
+package org.apache.submarine.server.rest.workbecnh;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -25,6 +25,9 @@ import org.apache.submarine.server.workbench.database.entity.SysDictEntity;
 import org.apache.submarine.server.workbench.database.entity.SysDictItemEntity;
 import org.apache.submarine.server.response.JsonResponse;
 import org.apache.submarine.server.response.JsonResponse.ListResult;
+import org.apache.submarine.server.rest.workbench.SysDictItemRestApi;
+import org.apache.submarine.server.rest.workbench.SysDictRestApi;
+import org.apache.submarine.server.rest.workbench.SystemRestApi;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
