+++ submarine-server/server-core/src/test/java/org/apache/submarine/server/workbench/database/utils/DictAnnotationTest.java
@@ -23,9 +23,9 @@ import com.google.gson.GsonBuilder;
 import com.google.gson.internal.LinkedTreeMap;
 import com.google.gson.reflect.TypeToken;
 import org.apache.submarine.server.response.DictAnnotation;
-import org.apache.submarine.server.workbench.rest.CommonDataTest;
-import org.apache.submarine.server.workbench.rest.SysUserRestApi;
 import org.apache.submarine.server.response.JsonResponse;
+import org.apache.submarine.server.rest.workbecnh.CommonDataTest;
+import org.apache.submarine.server.rest.workbench.SysUserRestApi;
 import org.junit.Test;
 
 import javax.ws.rs.core.Response;
